+++ src/main/java/com/workspaceit/dccpos/helper/TrackingIdGenerator.java
@@ -2,7 +2,7 @@ package com.workspaceit.dccpos.helper;
 
 public class TrackingIdGenerator {
     private static final String supplyPrefix = "SP";
-    public static String getShipmentTrackingId(int id){
+    public static String getShipmentTrackingId(long id){
         return supplyPrefix+String.valueOf(id);
     }
 }
